+++ @types/window.d.ts
@@ -1,4 +1,4 @@
-import type { TeamSubscription } from '../libs/@guardian/libs/src/logger/@types/logger';
+import type { ManageSubscription } from '../libs/@guardian/libs/src/logger/@types/logger';
 import type { Switches } from '../libs/@guardian/libs/src/switches/@types/Switches';
 import type { google } from '../libs/@guardian/atoms-rendering/src/ima';
 import type { ImaManager } from '../libs/@guardian/atoms-rendering/src/YoutubeAtomPlayer';
@@ -8,9 +8,10 @@ declare global {
 	interface Window {
 		guardian?: {
 			logger?: {
-				subscribeTo: TeamSubscription;
-				unsubscribeFrom: TeamSubscription;
+				subscribeTo: ManageSubscription;
+				unsubscribeFrom: ManageSubscription;
 				teams: () => string[];
+				subscriptions: () => string[];
 			};
 			config?: {
 				page?: {