+++ libs/@guardian/atoms-rendering/src/index.test.ts
@@ -3,10 +3,6 @@ import * as packageExports from './index';
 describe('The package', () => {
 	it('exports everything it did before', () => {
 		expect(Object.keys(packageExports).sort()).toEqual([
-			'AudioAtom',
-			'ChartAtom',
-			'ExplainerAtom',
-			'GuideAtom',
 			'InteractiveAtom',
 			'InteractiveLayoutAtom',
 			'KnowledgeQuizAtom',
@@ -15,7 +11,6 @@ describe('The package', () => {
 			'RecipeAtom',
 			'TimelineAtom',
 			'VideoAtom',
-			'YoutubeAtom',
 		]);
 	});
 });
@@ -23,8 +18,6 @@ describe('The package', () => {
 // test that type exports have not been removed.
 // won't catch new types but I don't know how we can?
 export type {
-	ChartAtomType,
-	GuideAtomType,
 	QandaAtomType,
 	SharingUrlsType,
 	TimelineEvent,