+++ libs/@guardian/atoms-rendering/src/index.ts
@@ -1,7 +1,3 @@
-export { AudioAtom } from './AudioAtom';
-export { ChartAtom } from './ChartAtom';
-export { ExplainerAtom } from './ExplainerAtom';
-export { GuideAtom } from './GuideAtom';
 export { InteractiveAtom } from './InteractiveAtom';
 export { InteractiveLayoutAtom } from './InteractiveLayoutAtom';
 export { KnowledgeQuizAtom } from './KnowledgeQuiz';
@@ -10,11 +6,8 @@ export { QandaAtom } from './QandaAtom';
 export { RecipeSchemaAtom as RecipeAtom } from './RecipeSchemaAtom';
 export { TimelineAtom } from './TimelineAtom';
 export { VideoAtom } from './VideoAtom';
-export { YoutubeAtom } from './YoutubeAtom';
 
 export type {
-	ChartAtomType,
-	GuideAtomType,
 	QandaAtomType,
 	SharingUrlsType,
 	TimelineEvent,