+++ libs/@guardian/libs/src/deprecated-exports.ts
@@ -5,6 +5,10 @@
  */
 
 import { Pillar } from './format/Pillar';
+import { type Subscription } from './logger/@types/logger';
 
 /** @deprecated Use `Pillar` instead. */
 export const ArticlePillar = Pillar;
+
+/** @deprecated Use `Subscription` instead. */
+export type TeamName = Subscription;