+++ libs/@guardian/libs/src/index.test.ts
@@ -19,6 +19,7 @@ describe('The package', () => {
 			'isBoolean',
 			'isNonNullable',
 			'isObject',
+			'isOneOf',
 			'isString',
 			'isUndefined',
 			'joinUrl',