+++ libs/@guardian/libs/src/performance/serialise.ts
@@ -1,17 +1,19 @@
 import { isNonNullable } from '../isNonNullable/isNonNullable';
 import { isString } from '../isString/isString';
-import { isTeam } from '../logger/teamStyles';
+import { isSubscription } from '../logger/subscriptions';
 import type { Detail } from './@types/measure';
 
 /** Serialisation is an implementation detail */
 const SEPARATOR = ':';
 
-export const serialise = ({ team, name, action }: Detail) =>
-	[team, name, action].filter(isNonNullable).join(SEPARATOR);
+export const serialise = ({ subscription, name, action }: Detail) =>
+	[subscription, name, action].filter(isNonNullable).join(SEPARATOR);
 
 export const deserialise = (id: string): Detail | undefined => {
-	const [team, name, action] = id.split(SEPARATOR);
-	return isString(team) && isTeam(team) && isString(name)
-		? { team, name, action }
+	const [subscription, name, action] = id.split(SEPARATOR);
+	return isString(subscription) &&
+		isSubscription(subscription) &&
+		isString(name)
+		? { subscription, name, action }
 		: undefined;
 };