+++ libs/@guardian/source-react-components-development-kitchen/src/toggle-switch/styles.ts
@@ -84,7 +84,7 @@ export const toggleStyles = (format?: ArticleFormat): SerializedStyles => {
 			top: 5px;
 			height: 11px;
 			width: 6px;
-			right: 8px;
+			right: 10px;
 			opacity: 0;
 			border-bottom: 2px solid ${success[400]};
 			border-right: 2px solid ${success[400]};